+++ cli/internal/oppath/oppath.go
@@ -1,0 +1,6 @@
+// Package oppath exposes functionality for resolving ops without absolute pathing.
+//
+// The idea of an op path is conceptually & functionally similar to an executable path
+// an executable path (typically controllable via a $PATH env var) allowing running
+// executables without specifying their absolute path.
+package oppath