+++ cli/ls.go
@@ -12,6 +12,8 @@ import (
 	"github.com/opctl/opctl/cli/internal/cliparamsatisfier"
 	"github.com/opctl/opctl/cli/internal/dataresolver"
 	"github.com/opctl/opctl/cli/internal/nodeprovider/local"
+	"github.com/opctl/opctl/cli/internal/oppath"
+	"github.com/opctl/opctl/sdks/go/model"
 	"github.com/opctl/opctl/sdks/go/opspec"
 )
 
@@ -37,24 +39,11 @@ func ls(
 		node,
 	)
 
-	_tabWriter := new(tabwriter.Writer)
-	defer _tabWriter.Flush()
-	_tabWriter.Init(os.Stdout, 0, 8, 1, '\t', 0)
-
-	fmt.Fprintln(_tabWriter, "REF\tDESCRIPTION")
-
-	dirHandle, err := dataResolver.Resolve(
+	opPath, err := oppath.Get(
 		ctx,
 		dirRef,
-		nil,
-	)
-	if err != nil {
-		return err
-	}
-
-	opsByPath, err := opspec.List(
-		ctx,
-		dirHandle,
+		dataResolver,
+		node,
 	)
 	if err != nil {
 		return err
@@ -65,33 +54,65 @@ func ls(
 		return err
 	}
 
-	for path, op := range opsByPath {
-		opRef := filepath.Join(dirHandle.Ref(), path)
-		if filepath.IsAbs(opRef) {
-			// make absolute paths relative
-			relOpRef, err := filepath.Rel(cwd, opRef)
-			if err != nil {
-				return err
+	_tabWriter := new(tabwriter.Writer)
+	defer _tabWriter.Flush()
+	_tabWriter.Init(os.Stdout, 0, 8, 1, '\t', 0)
+
+	fmt.Fprintln(_tabWriter, "REF\tDESCRIPTION")
+
+	for _, dirRef := range opPath {
+		dirHandle, err := dataResolver.Resolve(
+			ctx,
+			dirRef,
+			nil,
+			nil,
+		)
+		if err != nil {
+			return err
+		}
+
+		opsByPath, err := opspec.List(
+			ctx,
+			dirHandle,
+		)
+		if err != nil {
+			return err
+		}
+
+		for path, op := range opsByPath {
+			opRef := filepath.Join(dirHandle.Ref(), path)
+
+			if filepath.IsAbs(opRef) && !strings.HasPrefix(opRef, nodeConfig.DataDir) {
+				// make absolute paths relative
+				relOpRef, err := filepath.Rel(
+					cwd,
+					opRef,
+				)
+				if err != nil {
+					return err
+				}
+
+				opRef = strings.TrimPrefix(relOpRef, model.DotOpspecDirName+"/")
+			} else {
+				opRef = strings.TrimPrefix(path, "/")
 			}
 
-			opRef = strings.TrimPrefix(relOpRef, ".opspec/")
-		}
+			description := op.Description
+			if strings.TrimSpace(description) == "" {
+				description = "-"
+			}
 
-		description := op.Description
-		if strings.TrimSpace(description) == "" {
-			description = "-"
+			scanner := bufio.NewScanner(strings.NewReader(description))
+			if scanner.Scan() {
+				// first line of description, add the op ref
+				fmt.Fprintf(_tabWriter, "%v\t%v", opRef, scanner.Text())
+			}
+			for scanner.Scan() {
+				// subsequent lines, don't add the op ref but let the description span multiple lines
+				fmt.Fprintf(_tabWriter, "\n\t%v", scanner.Text())
+			}
+			fmt.Fprintln(_tabWriter)
 		}
-
-		scanner := bufio.NewScanner(strings.NewReader(description))
-		if scanner.Scan() {
-			// first line of description, add the op ref
-			fmt.Fprintf(_tabWriter, "%v\t%v", opRef, scanner.Text())
-		}
-		for scanner.Scan() {
-			// subsequent lines, don't add the op ref but let the description span multiple lines
-			fmt.Fprintf(_tabWriter, "\n\t%v", scanner.Text())
-		}
-		fmt.Fprintln(_tabWriter)
 	}
 
 	return nil