+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file in
 accordance with
 [![keepachangelog 1.0.0](https://img.shields.io/badge/keepachangelog-1.0.0-brightgreen.svg)](http://keepachangelog.com/en/1.0.0/)
 
+## [0.1.69-alpha.0] - 2025-03-23
+
+### Added
+
+- Support default ops within git repositories via remote conventional `.opctl` git repository
+- Support default `pullCreds` for ops via `opctl auth [...]`
+
 ## [0.1.68] - 2025-03-22
 
 ### Fixed